% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_meta.R
\name{BWB_RAIN_GAUGES}
\alias{BWB_RAIN_GAUGES}
\title{BWB_RAIN_GAUGES}
\usage{
BWB_RAIN_GAUGES(mdb = mdb_rain_meta())
}
\arguments{
\item{mdb}{path to MS Access Database Containing Meta Data}
}
\description{
BWB_RAIN_GAUGES
}
