% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_meta.R
\name{distanceToNeighbour}
\alias{distanceToNeighbour}
\title{Distance Matrix to Matrix of Neighbours}
\usage{
distanceToNeighbour(mdist)
}
\arguments{
\item{mdist}{distance matrix as returned by \code{\link{getGaugeDistances}}}
}
\value{
matrix of neighbour names with row names = gauge names and column
  names = \code{n1, n2, n<n.gauges - 1>}
}
\description{
Convert distance matrix to matrix of neighbours
}
