% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_files.R
\name{extractRunoffData}
\alias{extractRunoffData}
\title{Separation of Fixed Width Columns}
\usage{
extractRunoffData(runoff_raw, columnWidth = 8, columnNames = NULL, version = 1)
}
\arguments{
\item{runoff_raw}{vector of character representing the text lines containing
data in columns of fixed width}

\item{columnWidth}{number of characters in each column}

\item{columnNames}{column names to be given to the result data frame}

\item{version}{1: loop, 2: sapply}
}
\value{
The function returns a data frame with column names as given in 
  \code{columnNames}.
}
\description{
Separation of columns of 'columnWidth' characters width each
}
