% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_utils.R
\name{gaugeNamesShort}
\alias{gaugeNamesShort}
\title{Rain Gauge Long Name to Short Name}
\usage{
gaugeNamesShort(colnames, underscore.rm = FALSE)
}
\arguments{
\item{colnames}{Column names read from Excel sheet}

\item{underscore.rm}{if TRUE, underscores are removed}
}
\description{
Rain Gauge Long Name to Short Name
}
