% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cctv_from_xml_be.R
\name{get_overview_of_columns}
\alias{get_overview_of_columns}
\title{Get Overview of Columns in Data Frames}
\usage{
get_overview_of_columns(x)
}
\arguments{
\item{x}{list of data frames}
}
\value{
data frame with each row representing one data frame in \code{x} and
  with as many columns as there are distinct columns in all of the data 
  frames in \code{x}. In a row \emph{i} the data frame contains "x" in those 
  columns that are contained in the \emph{i}-th data frame in list \code{x}
  and an empty string otherwise.
}
\description{
Get Overview of Columns in Data Frames
}
\examples{
get_overview_of_columns(list(
  df1 = data.frame(a = 1, b = 2), 
  df2 = data.frame(b = 3, c = 4),
  df3 = data.frame(a = 5, c = 6)
))

}
