% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_pi.R
\name{get_PI_data_Halensee}
\alias{get_PI_data_Halensee}
\title{Get PI Data Halensee}
\usage{
get_PI_data_Halensee(xls.file, sheetName = "mydata")
}
\arguments{
\item{xls.file}{full path to MS Excel file containing PI data of Halensee}

\item{sheetName}{name of the cell range or worksheet to read. Default:
"mydata". There should be a cell range named "mydata" in the Excel file.
Create this named cell range by 1. selecting one non-empty cell within the
actual body of the table, 2. pressing Strg+A (select all), 3. writing
"mydata" into the name field (white input field, just above the cell area
and left of the "edit bar" labelled "fx"), and (iv) pressing Enter.}
}
\description{
Get PI Data Halensee
}
