% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wasserportal_stations_table.R
\name{get_wasserportal_stations_table}
\alias{get_wasserportal_stations_table}
\title{Wasserportal Berlin: get overview options for stations}
\usage{
get_wasserportal_stations_table(
  type = get_overview_options()$groundwater$level,
  url_wasserportal = wasserportal_base_url()
)
}
\arguments{
\item{type}{type of stations table to retrieve. Valid options defined in 
get_overview_options(), default: get_overview_options()$groundwater$level}

\item{url_wasserportal}{base url to Wasserportal berlin (default: 
wasserportal_base_url())}
}
\value{
data frame with metadata for
}
\description{
Wasserportal Berlin: get overview options for stations
}
\examples{
types <- kwb.read::get_overview_options()
str(types)
sw_l <- kwb.read::get_wasserportal_stations_table(type = types$surface_water$water_level)
str(sw_l)
}
