% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wasserportal_raw.R
\name{get_wasserportal_variables}
\alias{get_wasserportal_variables}
\title{Get Names and IDs of the Variables of wasserportal.berlin.de}
\usage{
get_wasserportal_variables(station = NULL)
}
\description{
Get Names and IDs of the Variables of wasserportal.berlin.de
}
