% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_data_high.R
\name{getAggregatedBwbRainDataFromExcelFiles}
\alias{getAggregatedBwbRainDataFromExcelFiles}
\title{Get Aggregated BWB Rain Data From Excel Files}
\usage{
getAggregatedBwbRainDataFromExcelFiles(rain.xls.files, tableName,
  aggregation.interval = NA, columnNames = NULL,
  use2007Driver = FALSE, dbg = FALSE)
}
\arguments{
\item{rain.xls.files}{vector of full paths to xls files}

\item{tableName}{name of sheet in Excel file(s) containing the rain data}

\item{aggregation.interval}{aggregation interval in seconds. Default: NA (=
no aggregation)}

\item{columnNames}{optional. Vector of names of columns (after renaming with 
\code{\link{gaugeNamesShort}}) to return}

\item{use2007Driver}{if TRUE, 2007 driver is used even if extension of
\emph{xls} is "xls"}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\value{
data frame with time column \emph{DateTime}, indicating the begin of
  the 5-minutes interval to which the rain height corresponds.
}
\description{
read BWB rain data from multiple XLS files with possible aggregation. 
Timestamps given in Local Time are converted to UTC+1 (= always winter time)
}
