% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_data_high.R
\name{getBwbRainDataFromExcelFiles}
\alias{getBwbRainDataFromExcelFiles}
\title{Get BWB Rain Data From Excel Files}
\usage{
getBwbRainDataFromExcelFiles(
  xlsPaths,
  tableName = "Niederschlag$",
  columnNames = NULL,
  use2007Driver = FALSE,
  dbg = FALSE
)
}
\arguments{
\item{xlsPaths}{(vector of) full path(s) to Excel file(s)}

\item{tableName}{sheet name, default: "[Niederschlag$]"}

\item{columnNames}{optional. Vector of names of columns (after renaming with 
\code{\link{gaugeNamesShort}}) to return}

\item{use2007Driver}{if TRUE, 2007 driver is used even if extension of
\emph{xls} is "xls"}

\item{dbg}{if \code{TRUE} (default), debug messages are shown}
}
\description{
Get BWB Rain Data From Excel Files
}
