% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_meta.R
\name{getGaugeDistances}
\alias{getGaugeDistances}
\title{Returns Distances Between Gauges}
\usage{
getGaugeDistances(gaugeInfo = getGaugeInfo(), gauges = c("NknI",
  "NknII", "ChbI", "BlnX", "BlnIX", "Wil", "Wila", "BlnV", "BlnXI", "Lbg",
  "Hlg", "ZhlIe", "KoepIf", "Kar", "SpaII", "BlnIV"), dbg = FALSE)
}
\arguments{
\item{gaugeInfo}{data frame as returned by \code{getGaugeInfo}}

\item{gauges}{vector of names of gauges for which distances are to be
calculated}

\item{dbg}{if \code{TRUE} (default), debug messages are shown}
}
\description{
Returns matrix containing relative distances between gauges
}
