% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_pi.R
\name{getPiColumnRenames}
\alias{getPiColumnRenames}
\title{Get List of Renamings for PI-Identifiers}
\usage{
getPiColumnRenames(mdb = mdb_rain_meta())
}
\arguments{
\item{mdb}{path to MS Access Database Containing Meta Data}
}
\value{
list of key = value pairs with the keys representing the
  PI-identifiers and the values representing the real (short) names
}
\description{
Get a list of assignements mapping PI identifiers to "real" (short) names
}
