% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senates_grab_sample_data_1.R
\name{hsGetAllGsData1}
\alias{hsGetAllGsData1}
\title{Get All Grab Sample Data (1)}
\usage{
hsGetAllGsData1(
  csvdir,
  sep = ";",
  dateFormat = "\%d.\%m.\%Y",
  dec = ",",
  data.block.number = NA,
  csvFiles = dir(csvdir, "*.csv", full.names = TRUE)
)
}
\arguments{
\item{csvdir}{absolute path to directory in which csv files to be read are 
located.}

\item{sep}{separator in csv file, e.g. ";" or ","}

\item{dateFormat}{date format specifier describing the format in which dates 
are represented in the csv file. Use placeholders , \code{"\%d"} (day), 
\code{"\%m"} (month), \code{"\%y"} (2-digit year), \code{"\%Y"} (4-digit
year) to describe the date format. \code{"\%d.\%m.\%Y"},
\code{"\%d/\%m/\%y"}, \code{"\%Y-\%m-\%d"} are examples for valid format
specifiers.}

\item{dec}{decimal character: "." or ","}

\item{data.block.number}{passed to \code{\link{hsGetGsData1}}, for a 
description see there.}

\item{csvFiles}{vector of full paths to csv files to be read. All files must
contain data for the same monitoring points, in the same order! Default:
all csv files in \emph{csvdir}}
}
\description{
get Senate's grab sample data (type 1) from all csv files in a directory and
return it in a "all-in-one-table" data frame
}
\seealso{
\code{\link{hsGetGsData1}}
}
