% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senates_grab_sample_data_2.R
\name{hsGsData2ToList}
\alias{hsGsData2ToList}
\title{Grab Sample Data 2 to List}
\usage{
hsGsData2ToList(gsData2, dbg = FALSE)
}
\arguments{
\item{gsData2}{grab sample data structure as returned by
\code{\link{hsGetGsData2}}}

\item{dbg}{if \code{TRUE} (default), debug messages are shown}
}
\description{
Transform grab sample data (type 2) as returned by \code{\link{hsGetGsData2}}
to "all-in-one-table"-format
}
\seealso{
\code{\link{hsGetGsData2}, \link{hsGsData1ToList}}
}
