% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neubrandenburg.R
\name{merge_Neubrandenburg_data}
\alias{merge_Neubrandenburg_data}
\title{Merge Neubrandenburg Data}
\usage{
merge_Neubrandenburg_data(allData, keep.all = FALSE)
}
\arguments{
\item{allData}{data frame}

\item{keep.all}{logical. If \code{TRUE} (the default is \code{FALSE}) all
columns from \code{allData} are contained in the result data frame, 
otherwise only the first column (timestamp).}
}
\description{
Merge Neubrandenburg data into one consistent data frame
}
