% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_utils.R
\name{niceStationNames}
\alias{niceStationNames}
\title{Nice Station Names}
\usage{
niceStationNames(stations)
}
\arguments{
\item{stations}{vector of character containing station (= gauge) names
to be cleaned}
}
\description{
Substitutions: Berlin -> Bln, Umlaut o -> oe, spaces removed
}
