% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_correction.R
\name{read_BWB_rain_correction}
\alias{read_BWB_rain_correction}
\title{Read Correction Data}
\usage{
read_BWB_rain_correction(files, type = (BWB_rain_correction_types())[1], ...)
}
\arguments{
\item{files}{paths to input files}

\item{type}{file type, must be one of the type strings returned by 
\code{kwb.read:::BWB_rain_correction_types}}

\item{\dots}{arguments passed to the read-function corresponding to the
given \code{type}}
}
\description{
Read BWB's rain correction data (different formats supported)
}
