% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_corr_wide.R
\name{read_BWB_rain_correction_wide_day}
\alias{read_BWB_rain_correction_wide_day}
\title{Read Correction "Wide" Format}
\usage{
read_BWB_rain_correction_wide_day(
  files,
  sep = ";",
  country = "de",
  transpose = TRUE,
  dbg = TRUE,
  date.format = c("\%d \%B \%Y")[1],
  encoding = ""
)
}
\arguments{
\item{files}{paths to rain correction files}

\item{sep}{column separator}

\item{country}{one of "de" (German) or "en" (English) according to the format
numeric strings are given in}

\item{transpose}{logical. If \code{TRUE} (default) data are transposed so 
that the different rain gauges appear in different columns}

\item{dbg}{logical. If \code{TRUE}, debug messages are shown.}

\item{date.format}{format used to convert the date string into a date object}

\item{encoding}{Encoding used when reading the CSV file. Best would be to use
"UTF-8" but MS Excel uses system default}
}
\description{
Read correction data from CSV file(s) in "wide" format with days in columns
}
