% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_corr_wide.R
\name{read_BWB_rain_correction_wide_gauge}
\alias{read_BWB_rain_correction_wide_gauge}
\title{Read Corrections in "wide" format}
\usage{
read_BWB_rain_correction_wide_gauge(
  files,
  sep = ";",
  country = "de",
  date.format = c("\%d.\%m.\%Y", "\%A \%d. \%B \%Y")[1],
  dbg = TRUE
)
}
\arguments{
\item{files}{paths to rain correction files}

\item{sep}{column separator}

\item{country}{one of "de" (German) or "en" (English) according to the format
numeric strings are given in}

\item{date.format}{format used to convert the date string into a date object}

\item{dbg}{logical. If \code{TRUE}, debug messages are shown.}
}
\description{
Read correction data from CSV file(s) in "wide" format with gauges in columns
}
