% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_corr_wide.R
\name{read_BWB_rain_correction_wide_gauge_from_pdf}
\alias{read_BWB_rain_correction_wide_gauge_from_pdf}
\title{Read Corrections in "Wide" Format}
\usage{
read_BWB_rain_correction_wide_gauge_from_pdf(file, sep = ";", ...)
}
\arguments{
\item{file}{path(s) to CSV file(s)}

\item{sep}{column separator}

\item{\dots}{passed to \code{kwb.read:::.extract_wide_gauge}}
}
\description{
Read correction data from CSV file(s) in "wide" format with gauges in 
columns. Version 2: As generated from PDF-file with "pdf to Excel"
}
