% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_corr_wide.R
\name{read_BWB_rain_correction_wide_gauge_group}
\alias{read_BWB_rain_correction_wide_gauge_group}
\title{Read Correction Data}
\usage{
read_BWB_rain_correction_wide_gauge_group(file,
  encoding = default_encoding(), ...)
}
\arguments{
\item{file}{path to rain correction data file}

\item{encoding}{file encoding, passed to \code{readLines}}

\item{\dots}{arguments passed to \code{kwb.read:::.extract_wide_gauge_group}}
}
\description{
Read Correction Data in "wide" format, grouped by groups of pumping stations
}
