% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cctv_from_xml_be.R
\name{read_cctv_from_xml_be}
\alias{read_cctv_from_xml_be}
\title{Read CCTV Data from XML File(s) in Belgian Standard Format}
\usage{
read_cctv_from_xml_be(files, as_is = FALSE, dbg = TRUE)
}
\arguments{
\item{files}{(vector of) path(s) to XML file(s)}

\item{as_is}{if \code{TRUE} the column names in the table of observations are 
more or less kept as they are in the xml files, otherwise, they are renamed
to the names that are defined in the EN 13508.2 standard. These names are
e.g. required by the RERAU evaluation functions in \code{kwb.rerau}. The
default is \code{FALSE}.}

\item{dbg}{if \code{TRUE} (default), debug messages are shown}
}
\description{
Read CCTV Data from XML File(s) in Belgian Standard Format
}
