% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neubrandenburg.R
\name{read_Neubrandenburg_Data}
\alias{read_Neubrandenburg_Data}
\title{Join Data from Multiple Excel Files}
\usage{
read_Neubrandenburg_Data(xls.dir, area.Gruendach.m2, area.Kiesdach.m2,
  pattern = "\\\\.xlsx$")
}
\arguments{
\item{xls.dir}{path to directory containing Excel files to be read}

\item{area.Gruendach.m2}{area of green roof in square metres. Used to 
convert between litres and litres per square metres (mm) as given in the
result data frame}

\item{area.Kiesdach.m2}{area of pebble roof in square metres. Used to 
convert between litres and litres per square metres (mm) as given in the
result data frame}

\item{pattern}{pattern matching the names of files in \code{xls.dir} to be 
imported}
}
\description{
Join Data from Multiple Excel Files
}
