% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neubrandenburg.R
\name{read_Neubrandenburg_Data_from_Excel}
\alias{read_Neubrandenburg_Data_from_Excel}
\title{Read Neubrandenburg Data}
\usage{
read_Neubrandenburg_Data_from_Excel(xls)
}
\arguments{
\item{xls}{full path to MS Excel file containing Neubrandenburg rain gauge
and green roof discharge data}
}
\value{
data frame with columns \emph{DateTime} (representing Local Normal
  Time but with attribute tzone = "UTC"), \emph{NSB}: Niederschlag,
  \emph{KFS}: Kiesdach (Kippe, Fallrohr, Stein), \emph{KFZ}: Gruendach
  (Kippe, Fallrohr, Zinko)
}
\description{
Read Neubrandenburg data from Excel file
}
