% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neubrandenburg.R
\name{read_Neubrandenburg_Data_from_Excel_files}
\alias{read_Neubrandenburg_Data_from_Excel_files}
\title{Read Neubrandenburg Data from Excel Files}
\usage{
read_Neubrandenburg_Data_from_Excel_files(xls.files)
}
\arguments{
\item{xls.files}{vector of (full paths to) MS Excel files containing
Neubrandenburg data}
}
\description{
Read and join Neubrandenburg data from multiple Excel files
}
