% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_files.R
\name{read_qin_file}
\alias{read_qin_file}
\title{Read "qin"-File}
\usage{
read_qin_file(infile, columnWidth = 8, n = -1, headerSize = 528,
  dbg = TRUE)
}
\arguments{
\item{infile}{full path to input file.}

\item{columnWidth}{width of the data columns. It is assumed that all data
columns have the same width!}

\item{n}{number of (data) rows to read. Set to -1 if all rows are to be read.
Default: -1}

\item{headerSize}{number of header rows preceeding the data rows in
\emph{infile}}

\item{dbg}{if TRUE, debug messages are shown, else not}
}
\description{
Read data, extract starttime, timestep, manhole IDs and runoff data
}
