% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_data_high.R
\name{read_rain}
\alias{read_rain}
\title{Read Rain Data from BWB Excel Files}
\usage{
read_rain(
  directory = kwb.utils::resolve("RAIN_DIR", dictionary),
  pattern = kwb.utils::resolve("RAIN_XLS_PATTERN", dictionary),
  tableName = kwb.utils::resolve("RAIN_XLS_TABLE", dictionary),
  dictionary = kwb.utils::selectElements(settings, "dictionary"),
  rain.xls.files = dir(directory, pattern, full.names = TRUE),
  aggregation.interval = kwb.utils::selectElements(settings,
    "rain.aggregation.interval"),
  settings = NULL,
  use2007Driver = TRUE
)
}
\arguments{
\item{directory}{directory containing excel files from which rain data are
read. Default: \code{kwb.utils::resolve("RAIN_DIR", dictionary)}}

\item{pattern}{file name pattern matching names of files to be read. Default:
\code{kwb.utils::resolve("RAIN_XLS_PATTERN", dictionary)}}

\item{tableName}{name of sheet in the excel files containing the rain data.
Default: \code{kwb.utils::resolve("RAIN_XLS_TABLE", dictionary)}}

\item{dictionary}{list of key/value pairs that can be used as a dictionary to
look up the following keywords (if \emph{directory}, \emph{pattern}, 
\emph{tableName} are not given): RAIN_DIR, RAIN_XLS_PATTERN, RAIN_XLS_TABLE}

\item{rain.xls.files}{full paths to xls or xlsx files to be read. If given,
these paths are used. Otherwise all files in \emph{directory} matching the
filename \emph{pattern} are used.}

\item{aggregation.interval}{aggregation interval in seconds. Default: 
settings$rain.aggregation.interval}

\item{settings}{list that may contain arguments to this function in elements
\code{rain.aggregation.interval} or \code{dictionary}}

\item{use2007Driver}{passed to
\code{\link{getAggregatedBwbRainDataFromExcelFiles}}}
}
\value{
data frame with column \code{DateTime} containing timestamps in UTC 
  timezone but representing Berlin normal (= winter) time.
}
\description{
Read Rain Data from BWB Excel Files
}
