% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senates_grab_sample_data_new.R
\name{read_SenatesGrabSampleCSV_long}
\alias{read_SenatesGrabSampleCSV_long}
\title{Read Senates' Grab Sample File}
\usage{
read_SenatesGrabSampleCSV_long(csv, sep = kwb.utils::guessSeparator(csv),
  dateFormat, dec = ifelse(sep == ",", ".", ","), dbg = FALSE)
}
\arguments{
\item{csv}{path to CSV file}

\item{sep}{column separator}

\item{dateFormat}{date format string, passed to
\code{kwb.read:::.removeNonDataRows} and
\code{kwb.read:::.convertDataTypes}}

\item{dec}{decimal character}

\item{dbg}{logical. If \code{TRUE}, debug messages are shown.}
}
\description{
Read Senates' grab sample file ('long' format). For a format description see 
\code{\link{hsGetGsData1}}
}
