% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senates_grab_sample_data_new.R
\name{read_SenatesGrabSampleCSV_wide}
\alias{read_SenatesGrabSampleCSV_wide}
\title{Read Senates' Grab Sample File}
\usage{
read_SenatesGrabSampleCSV_wide(csv, sep, dateFormat, country = "de",
  doConversion = TRUE, dbg = TRUE, oldversion = FALSE)
}
\arguments{
\item{csv}{path to CSV file}

\item{sep}{column separator}

\item{dateFormat}{date format string, passed to
\code{kwb.read:::.doConversion} if \code{doConversion} is \code{TRUE}}

\item{country}{one of "de" (German) or "en" (English) according to the format
numeric strings are given in}

\item{doConversion}{logical. If \code{TRUE}, the function
\code{kwb.read:::.doConversion} is called on the result}

\item{dbg}{logical. If \code{TRUE}, debug messages are shown.}

\item{oldversion}{passed to \code{kwb.read:::.doConversion} if
\code{doConversion} is \code{TRUE}}
}
\description{
Read Senates' grab sample file ('wide' format). For a format description see 
\code{\link{hsGetGsData2}}
}
