% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senates_grab_sample_data_new.R
\name{read_SenatesGrabSampleCsvFiles_long}
\alias{read_SenatesGrabSampleCsvFiles_long}
\title{Read Senates' Grab Sample Files}
\usage{
read_SenatesGrabSampleCsvFiles_long(csvFiles, ...)
}
\arguments{
\item{csvFiles}{full paths to CSV files to be read}

\item{\dots}{arguments passed to \code{\link{read_SenatesGrabSampleCSV_long}}}
}
\description{
read Senates' grab sample files ('long' format). For a format description see
\code{\link{hsGetGsData1}}
}
