% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senates_grab_sample_data_new.R
\name{read_SenatesGrabSampleCsvFiles_wide}
\alias{read_SenatesGrabSampleCsvFiles_wide}
\title{Read Senates' Grab Sample Files}
\usage{
read_SenatesGrabSampleCsvFiles_wide(
  csvs,
  sep,
  dateFormat,
  country = "de",
  dbg = TRUE,
  oldversion = FALSE
)
}
\arguments{
\item{csvs}{vector of paths to CSV files}

\item{sep}{column separator}

\item{dateFormat}{date format string, passed to
\code{\link{read_SenatesGrabSampleCSV_wide}} and
\code{kwb.read:::.doConversion}}

\item{country}{one of "de" (German) or "en" (English) according to the format
numeric strings are given in}

\item{dbg}{logical. If \code{TRUE}, debug messages are shown.}

\item{oldversion}{passed to \code{kwb.read:::.doConversion}}
}
\description{
Read Senates' grab sample files ('wide' format). For a format description see
\code{\link{hsGetGsData2}}
}
