% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_as_path_value.R
\name{read_xml_as_path_value}
\alias{read_xml_as_path_value}
\title{Read XML File as Pairs of Paths and Values}
\usage{
read_xml_as_path_value(xml, dbg = TRUE, max_length = 100)
}
\arguments{
\item{xml}{path, url or literal xml or anything else that is accepted by
\code{\link[xml2]{read_xml}} as argument \code{x}}

\item{dbg}{if \code{TRUE}, debug messages are shown}

\item{max_length}{maximum number of characters reserved for printing the
path to the folder of the \code{xml} file in the debug message}
}
\value{
data frame with the full paths to the XML elements in the first
  and the text values of the XML elements in the second column.
}
\description{
Read XML File as Pairs of Paths and Values
}
\examples{
url <- "https://www.w3schools.com/xml/note.xml"

# Original XML content
kwb.utils::catLines(readLines(url, warn = FALSE))

# Interpretation as Paths and Values
kwb.read:::read_xml_as_path_value(url)

}
