% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_data_high.R
\name{readAllBwbCorrections}
\alias{readAllBwbCorrections}
\title{Read and Merge BWB Correction Files}
\usage{
readAllBwbCorrections(dir.in, args.general, args, dbg = TRUE)
}
\arguments{
\item{dir.in}{full path to input directory where to look for CSV files to
import}

\item{args.general}{list of general argument settings to be used when calling
\code{\link{read_BWB_rain_correction}}. For available argument names, see 
the help pages of the sub-functions 
\code{\link{read_BWB_rain_correction_long}}, 
\code{\link{read_BWB_rain_correction_wide_day}}, 
\code{\link{read_BWB_rain_correction_wide_gauge}}.}

\item{args}{list of argument settings for each file. The elements in the list
are expected to be the names of files in \code{dir.in}.}

\item{dbg}{if \code{TRUE} (default), debug messages are shown}
}
\description{
Read and Merge BWB Correction Files
}
