% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_data_high.R
\name{readAllBwbCumulativeHeights}
\alias{readAllBwbCumulativeHeights}
\title{Read and Combine Rain Heights}
\usage{
readAllBwbCumulativeHeights(files.data, ..., dbg = TRUE)
}
\arguments{
\item{files.data}{character vector containing the full paths to the CSV files
to be read}

\item{\dots}{arguments passed to \code{\link{readBwbRainFromCsv2}}}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\description{
Read rain height files with \code{\link{readBwbRainFromCsv2}}, combine them
by rows, remove duplicates and order by timestamp
}
