% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_data_high.R
\name{readAllBwbSignals}
\alias{readAllBwbSignals}
\title{Read and Merge BWB Rain Data Files}
\usage{
readAllBwbSignals(files.data, date.format = "\%Y-\%m-\%d", dbg = TRUE)
}
\arguments{
\item{files.data}{character vector containing the full paths to the CSV files
to be read}

\item{date.format}{format used to interpret the date string. See 
\code{\link[base]{format.POSIXct}}}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\description{
Read and Merge BWB rain data files with \code{\link{readBwbRainFromCsv2}}
}
