% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_correction.R
\name{readBwbCorrFromCsv}
\alias{readBwbCorrFromCsv}
\title{Read BWB Rain Correction Data from CSV File}
\usage{
readBwbCorrFromCsv(file, sep = ",", country = "de", format = "\%d.\%m.\%Y")
}
\arguments{
\item{file}{path to CSV file}

\item{sep}{column separator}

\item{country}{contry code specifying decimal and thousands separators, one}

\item{format}{date format, passed to \code{\link{as.Date}}}
}
\description{
Read BWB rain correction data from a CSV file that has been created by saving
the corresponding Excel file as CSV
}
