% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_correction.R
\name{readBwbCorrFromExcel}
\alias{readBwbCorrFromExcel}
\title{Read BWB Rain Correction Data from Excel File}
\usage{
readBwbCorrFromExcel(
  file,
  tblCorr = "Bericht 1$",
  tblGaug = "gaugeNames",
  dbg = TRUE
)
}
\arguments{
\item{file}{full path to Excel file}

\item{tblCorr}{name of sheet containing correction data. Default: 
"Bericht 1$"}

\item{tblGaug}{name of sheet containing gauge Names. Default: "gaugeNames$"}

\item{dbg}{logical. If \code{TRUE}, debug messages are shown.}
}
\description{
Read BWB Rain Correction Data from Excel File
}
