% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_correction.R
\name{readBwbRainCorrection}
\alias{readBwbRainCorrection}
\title{Read BWB Rain Correction Data}
\usage{
readBwbRainCorrection(file, ..., zerolines.rm = TRUE, dbg = FALSE)
}
\arguments{
\item{file}{full path to Excel file or CSV file}

\item{\dots}{arguments given to readBwbCorrFromExcel or readBwbCorrFromCSV}

\item{zerolines.rm}{logical. If \code{TRUE}, rows in which the sum of 
absolute values is zero, are removed}

\item{dbg}{logical. If \code{TRUE}, debug messages are shown.}
}
\description{
read BWB rain correction data from Excel file or CSV file
}
