% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_data.R
\name{readBwbRainData}
\alias{readBwbRainData}
\title{Read BWB Rain Data}
\usage{
readBwbRainData(
  file,
  tableName = "Niederschlag$",
  columnNames = NULL,
  toUTC = TRUE,
  toNormaltime = FALSE,
  dbg = TRUE,
  use2007Driver = kwb.db::isExcel2007File(file),
  ...
)
}
\arguments{
\item{file}{full path to Excel file}

\item{tableName}{sheet name, default: "[Niederschlag$]"}

\item{columnNames}{optional. Vector of names of columns (after renaming with 
\code{\link{gaugeNamesShort}}) to return}

\item{toUTC}{default: TRUE}

\item{toNormaltime}{if TRUE, a column \emph{tBeg_WT} is created containing
the begin of the time interval in Normal time (winter time, UTC+01).
default: FALSE}

\item{dbg}{default: TRUE}

\item{use2007Driver}{if TRUE, 2007 driver is used even if extension of
\emph{file} is ".xls"}

\item{\dots}{additional parameters passed to \code{\link{readBwbRainFromCsv}}}
}
\description{
Read BWB Rain Data
}
