% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_data.R
\name{readBwbRainFromCsv}
\alias{readBwbRainFromCsv}
\title{Read BWB Rain Data from CSV File}
\usage{
readBwbRainFromCsv(
  file,
  sep = ";",
  dec = ",",
  format = "\%d.\%m.\%y \%H:\%M:\%S"
)
}
\arguments{
\item{file}{path to CSV file}

\item{sep}{column separator}

\item{dec}{decimal character}

\item{format}{date format string, passed to \code{\link{as.POSIXct}}}
}
\description{
read BWB rain data from a CSV file that has been created by saving the 
corresponding Excel file as CSV
}
