% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_data_high.R
\name{readYearlyRainHeightsFromOneCsvFile}
\alias{readYearlyRainHeightsFromOneCsvFile}
\title{Read Yearly Rain Heights From One Csv File}
\usage{
readYearlyRainHeightsFromOneCsvFile(csv, sep = ",",
  dateformat = "\%d/\%m/\%Y")
}
\arguments{
\item{csv}{full path to csv file}

\item{sep}{column separator. default: comma ","}

\item{dateformat}{date format, default: "\%d/\%m/\%Y"}
}
\value{
data frame with columns...
}
\description{
reads a CSV file containing daily rain heights as provided by BWB (Mario 
Grunwald). Example filename: "Niederschlaege_1994__BERICHT.csv". The files 
are expected to contain a header of three rows (first row: DATUM/station 
names, second row: variable name ["Regenhoehe" or "Regendauer"], third row: 
unit ["in mm" or "in mm"])
}
