% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwb_rain_utils.R
\name{toGaugeInfo}
\alias{toGaugeInfo}
\title{BWB Rain Data Column Name to Gauge Info Table}
\usage{
toGaugeInfo(x)
}
\arguments{
\item{x}{vector of column names of the form "01.02 <gauge> - Regenmesser"}
}
\description{
Convert a vector of BWB rain data column names "\code{id} \code{name} -
Regenmesser" to a data frame with columns \code{id} and \code{<name>}
}
